t3310: extract common no_notes_merge_left()
We had many statements which were duplicated. Extract and replace these
duplicate statements with no_notes_merge_left().

While we're at it, replace the test_might_fail(), which should only be
used on git commands, with a compound command that always returns 0,
even if the underlying ls fails.

Also, remove the redirection from stderr to /dev/null. This is because
the test scripts automatically suppress output already. Otherwise, if a
developer asks for verbose output via the `-v` flag, the stderr output
may be useful for debugging.

Signed-off-by: Denton Liu <liu.denton@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>